Clear a field before filling it (#46)
* Clear fields before filling them

During a test, when the test fills a field that already has content in
it (e.g. `Name: Edward`, `name_field.fill("Paul")`) the expectation is
that the field's value would be just the new value. However, it's both
(e.g. `Name: EdwardPaul`). This comes up especially in tests that edit
an existing model via a form.

This change clears a field before filling it so that values are
expected.

* Add append method

Just in case someone actually wants that feature, and it's so easy to
add.

* Fill, append, clear documentation